automation/functions/requirements.txt
azure-functions
azure-identity
azure-mgmt-applicationinsights
azure-mgmt-monitor
requests
